Move static funcs to utils file p2
