+++ checkov/terraform/__init__.py
@@ -1,2 +1,3 @@
 from checkov.terraform.checks.resource import *  # noqa
 from checkov.terraform.checks.data import *  # noqa
+from checkov.terraform.modules import *  # noqa
